requirements.txt
breeze-connect>=1.0.0
flask>=2.0.0
gunicorn>=21.0.0
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
python-dateutil>=2.8.0
APScheduler>=3.10.0
